Add option to not open the socials link on a new page, i.e. for an extra email page
I've added a ``nonewpage`` parameter to social links to avoid popping them open on a new page. This is useful, for example, for adding an email page that is resistant to scraping.

As an example:
``
[[params.social]]
  name = "email"
  url  = "/email"
  nonewpage = true
``

Where /email.md is a page that says something like ``domain at domain dot tld``.